+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.145.0] - Not released
+### Changed
+- Limit feed entries number to 60 (configurable via `CONFIG.feed.maxEntries`) to
+  prevent the feed from growing indefinitely on long-running sessions.
 ### Fixed
 - Prevent closing lightbox on wheel scroll if the image is zoomed in.
 - Fix data parsing on Calendar pages.